1.2.0-alpha0: Big fonts, refactoring (#11)
- Initial "big fonts" support
  - Primarily designed for running at native 4k resolution
  - The larger fonts requires the use of custom 2x upscaled ui skins
  - The "big fonts" mode is enabled if the active ui skin folder
    contains a file named 'use_zeal_big_fonts.txt' in it (can be empty)
    - Note: Big fonts are activated at boot and requires a client
      re-start to transition
  - Quit and restart client after using /load or /reloadskin to new mode

Infrastructure:
- Updated changelog and readme with 1.1 notes
- Refactored headers to reduce compile time and make dependencies more obvious
  - Eliminated catch-all "framework.h" and added module specific headers
- Refactored the dll to eliminate the separate Zeal lifetime thread and
  instantiates Zeal on the primary calling attach thread
  - Simplifies the code; may help isolate the client's boot heap corruption
- Refactored the ZealService to simplify and make it more modular
  - Moved auxiliary functions (binds, config check) to more relevant modules
- New UISkin class to centralize zeal file resource access and handles big fonts
- Zeal now uses a hard-coded whitelist of required xml files and no longer tries
  to load other random xmls from the zeal/ folder
- Updated xml and resource file path accesses to use std::filesystem::path
- Refactored GameClass and Display classes to add methods and expand fields

+++ Zeal/alarm.h
@@ -1,6 +1,4 @@
 #pragma once
-#include "hook_wrapper.h"
-#include "memory.h"
 
 class Alarm {
  public:
@@ -11,7 +9,7 @@ class Alarm {
 
  private:
   void callback_main();
-  ULONGLONG start_time = 0;
+  unsigned long long start_time = 0;
   double duration = 0;
   bool enabled = false;
 };
